+++ NEWS.md
@@ -1,7 +1,9 @@
 # A update version visa 1.0.0 [2025-03-15]
 * Main changes: 
-* adding cm.rbd3, find.bestBands, and visualization with plot3D
-* modify defined database to library in spectra-class
+* adding cm.rbd3, find.bestBands, with visualization via plot3D.
+* modify defined database to library in spectra-class.
+* Renamed function `ggplot.cm` to `plot.2dcm`.
+* two plot.3dcm functions
 * major version number update to 1.0.0
 
 # First release visa 0.1.0  [2021-04-20]